requirements.txt
asyncio
requests
aiogram
example
flask
python_dotenv
huggingface
huggingface_hub